+++ backend/src/main/java/com/woowacourse/gongseek/vote/application/dto/VoteResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.vote.presentation.dto;
+package com.woowacourse.gongseek.vote.application.dto;
 
 import com.woowacourse.gongseek.vote.domain.Vote;
 import com.woowacourse.gongseek.vote.domain.repository.dto.VoteItemDto;
